xmore_news_reliable/requirements.txt
# Xmore Reliable News Acquisition Layer
# Python 3.10+
# Install: pip install -r requirements.txt

# ---------------------------------------------------------------------------
# Gmail API (OAuth2)
# ---------------------------------------------------------------------------
google-api-python-client>=2.115.0
google-auth>=2.27.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# ---------------------------------------------------------------------------
# RSS / Feed parsing
# ---------------------------------------------------------------------------
feedparser>=6.0.11

# ---------------------------------------------------------------------------
# HTML parsing (email body + page watcher)
# ---------------------------------------------------------------------------
beautifulsoup4>=4.12.3
lxml>=5.1.0               # faster HTML parser backend for BeautifulSoup

# ---------------------------------------------------------------------------
# HTTP client (PDF downloader + page watcher)
# ---------------------------------------------------------------------------
requests>=2.31.0

# ---------------------------------------------------------------------------
# PDF text extraction (install at least one)
# ---------------------------------------------------------------------------
pdfplumber>=0.10.3        # primary — best for structured financial PDFs
# PyMuPDF>=1.23.0         # fallback — faster, handles more corrupt PDFs
#                         # uncomment if needed: pip install PyMuPDF

# ---------------------------------------------------------------------------
# Configuration
# ---------------------------------------------------------------------------
python-dotenv>=1.0.1
